+++ benchmarks/FNO/Rotor37/prepare_rotor37.py
@@ -1,5 +1,5 @@
-from plaid.containers.sample import Sample
-from plaid.problem_definition import ProblemDefinition
+from plaid import Sample
+from plaid import ProblemDefinition
 from Muscat.Bridges.CGNSBridge import CGNSToMesh
 import numpy as np
 from Muscat.Containers.Filters.FilterObjects import ElementFilter