+++ benchmarks/FNO/VKI-LS59/prepare_vki.py
@@ -1,5 +1,5 @@
-from plaid.problem_definition import ProblemDefinition
-from plaid.containers.sample import Sample
+from plaid import ProblemDefinition
+from plaid import Sample
 from Muscat.Bridges.CGNSBridge import MeshToCGNS
 from Muscat.Containers.ConstantRectilinearMeshTools import CreateConstantRectilinearMesh
 from Muscat.Containers.MeshTetrahedrization import Tetrahedrization