+++ benchmarks/Vi-Transf/src/data/loader/bridges/multiscale_sample_to_geometric.py
@@ -3,8 +3,8 @@ import torch
 from sklearn.neighbors import KDTree
 from torch_geometric.data import Data
 
-from plaid.containers.sample import Sample
-from plaid.problem_definition import ProblemDefinition
+from plaid import Sample
+from plaid import ProblemDefinition
 
 from .utils import faces_to_edges
 
@@ -29,7 +29,7 @@ def multiscale_sample_to_geometric(
     """Converts a Plaid sample to PytorchGeometric Data object.
 
     Args:
-        sample (plaid.containers.sample.Sample): data sample
+        sample (plaid.Sample): data sample
 
     Returns:
         Data: the converted data sample