+++ CHANGELOG.md
@@ -12,11 +12,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - (docs) update logos
+- (imports) add imports of `Sample`, `Dataset` from `plaid` and `plaid.containers` and `ProblemDefinition` from `plaid`
+- (dataset.py) add optional `ids` argument to `from_list_of_samples`
+- (sample) add summarize and check_completeness functions
+- (dataset) add summarize_features and check_feature_completeness functions
 - (Hugging Face bridge) add datasetdict conversion, and simple function for plaid sample init from hf sample
 - (pipelines/plaid_blocks.py) add column transformer inverse_transform
+
+### Changed
+
 - Move to Muscat=2.5.0 (for tests and examples support)
 - Update repo configuration (actions: rely more on pypi dependencies, action versions)
 - Rename types to remove `Type` from name of types: https://github.com/PLAID-lib/plaid/pull/164
+
+### Fixes
+
+- (plaid/examples) fix circular imports
 
 ### Removed